requirements.txt
scikit-learn==1.8.0
xgboost==2.1.4
numpy==2.2.5
pandas==2.2.3
streamlit==1.51.0
joblib==1.5.3
fastapi==0.117.1
fastapi-cli==0.0.13
flake8==7.3.0
seaborn==0.13.2
matplotlib==3.10.5
pyyaml==6.0.2
plotly==5.18.0
uvicorn==0.37.0
mlflow==3.8.1
boto3>=1.34.0
pydantic==2.8.2
python-multipart==0.0.9
requests==2.32.4
black==24.10.0
isort==4.2.5
pytest==9.0.2
#pylint==2.17.5
